added db context tables and config

+++ KH.Domain/Commons/HistoryTrackerEntity.cs
@@ -7,7 +7,7 @@ namespace KH.Domain.Commons
         [Column(Order = 103)]
         public Nullable<DateTime> UpdatedDate { get; set; }
         [Column(Order = 104)]
-        public Nullable<int> UpdatedById { get; set; }
+        public Nullable<long> UpdatedById { get; set; }
     }
 
 }
