+++ KH.Domain/Commons/TrackerEntity.cs
@@ -9,20 +9,20 @@ namespace KH.Domain.Commons
 		[Column(Order = 101)]
 		public DateTime CreatedDate { get; set; }
 		[Column(Order = 102)]
-		public Nullable<int> CreatedById { get; set; }
+		public Nullable<long> CreatedById { get; set; }
         //public User? CreatedByUser { get; set; }
         [Column(Order = 103)]
 		public Nullable<DateTime> UpdatedDate { get; set; }
         //public User? UpdatedByUser { get; set; }
 
         [Column(Order = 104)]
-		public Nullable<int> UpdatedById { get; set; }
+		public Nullable<long> UpdatedById { get; set; }
 		[Column(Order = 105)]
 		public bool IsDeleted { get; set; } = false;
 		[Column(Order = 106)]
 		public Nullable<DateTime> DeletedDate { get; set; }
 		[Column(Order = 107)]
-		public Nullable<int> DeletedById { get; set; }
+		public Nullable<long> DeletedById { get; set; }
         //public User? DeletedByUser { get; set; }
 
     }
