+++ KH.Domain/Entities/Role.cs
@@ -3,7 +3,7 @@ namespace KH.Domain.Entities
 {
     public class Role : BasicEntity
     {
-        public int? ReportToRoleId { get; set; }
+        public long? ReportToRoleId { get; set; }
         public Role? ReportToRole { get; set; }
 
     }
