+++ KH.Domain/Entities/RoleFunction.cs
@@ -1,12 +1,12 @@
 
 namespace KH.Domain.Entities
 {
-    public class RoleFunction : TrackerEntity
-    {
-        public int SystemFunctionId { get; set; }
-        public SystemFunction SystemFunction { get; set; }
+  public class RoleFunction : TrackerEntity
+  {
+    public long SystemFunctionId { get; set; }
+    public SystemFunction SystemFunction { get; set; }
 
-		public int RoleId { get; set; }
-		public Role Role { get; set; }
-	}
+    public long RoleId { get; set; }
+    public Role Role { get; set; }
+  }
 }
