+++ KH.Domain/Entities/UserDepartment.cs
@@ -3,10 +3,10 @@ namespace KH.Domain.Entities
 {
     public class UserDepartment : TrackerEntity
     {
-        public int UserId { get; set; }
+        public long UserId { get; set; }
         public User? User { get; set; }
 
-        public int DepartmentId { get; set; }
+        public long DepartmentId { get; set; }
         public Department? Department { get; set; }
     }
 }
