+++ KH.Domain/Entities/UserRole.cs
@@ -3,16 +3,16 @@ using System.ComponentModel.DataAnnotations.Schema;
 
 namespace KH.Domain.Entities
 {
-    public class UserRole : TrackerEntity
-    {
-        public int UserId { get; set; }
-        public User? User { get; set; }
+  public class UserRole : TrackerEntity
+  {
+    public long UserId { get; set; }
+    public User? User { get; set; }
 
-        public int RoleId { get; set; }
-        public Role? Role { get; set; }
+    public long RoleId { get; set; }
+    public Role? Role { get; set; }
 
-        [NotMapped]
-        public List<RoleFunction> RoleFunctions { get; set; } = new List<RoleFunction>();
+    [NotMapped]
+    public List<RoleFunction> RoleFunctions { get; set; } = new List<RoleFunction>();
 
-    }
+  }
 }
