+++ KH.Domain/Entities/lookups/Group.cs
@@ -2,7 +2,7 @@ namespace KH.Domain.Entities.lookups
 {
     public class Group : BasicEntity
     {
-        public int? TicketCategoryId { get; set; }
+        public long? TicketCategoryId { get; set; }
     }
 
 }
